+++ internal/log/middleware_test.go
@@ -253,18 +253,18 @@ func BenchmarkDataRace(b *testing.B) {
 
 func TestForwardedAddrHandler(t *testing.T) {
 	out := &bytes.Buffer{}
-	r := &http.Request{
-		Header: http.Header{
-			"X-Forwarded-For": []string{"client", "proxy1", "proxy2"},
-		},
-	}
+
+	r := httptest.NewRequest(http.MethodGet, "/", nil)
+
+	r.Header.Set("X-Forwarded-For", "proxy1,proxy2,proxy3")
+
 	h := ForwardedAddrHandler("fwd_ip")(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		l := FromRequest(r)
 		l.Log().Msg("")
 	}))
 	h = NewHandler(zerolog.New(out))(h)
 	h.ServeHTTP(nil, r)
-	if want, got := `{"fwd_ip":"client"}`+"\n", decodeIfBinary(out); want != got {
+	if want, got := `{"fwd_ip":["proxy1","proxy2","proxy3"]}`+"\n", decodeIfBinary(out); want != got {
 		t.Errorf("Invalid log output, got: %s, want: %s", got, want)
 	}
 }